backend/requirements.txt
azure-functions

streamlit==1.24.0
openai==0.27.8
matplotlib==3.6.3
plotly==5.12.0
scipy==1.10.0
scikit-learn==1.2.0
transformers==4.30.0
python-dotenv==1.0.0
azure-ai-formrecognizer==3.2.0
azure-storage-blob==12.14.1
azure-identity==1.12.0
azure-ai-contentsafety==1.0.0b1
requests==2.31.0
tiktoken==0.2.0
azure-storage-queue==12.5.0
langchain==0.0.274
beautifulsoup4==4.12.0
fake-useragent==1.1.3
chardet==5.1.0
--extra-index-url https://pkgs.dev.azure.com/azure-sdk/public/_packaging/azure-sdk-for-python/pypi/simple/
azure-search-documents==11.4.0b8
opencensus-ext-azure==1.1.9
pandas==1.5.1
python-docx==0.8.11
gpt4all==1.0.8
